backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bip-utils==2.9.3
bitarray==2.9.2
blis==1.0.1
catalogue==2.0.10
cbor2==5.6.4
cdp-sdk==0.11.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
cloudpathlib==0.20.0
coincurve==20.0.0
confection==0.1.5
crcmod==1.7
cryptography==43.0.1
cymem==2.0.10
cytoolz==0.12.3
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
ecdsa==0.19.0
ed25519-blake2b==1.4.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
hexbytes==1.2.1
hiredis==3.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
llama-index-core==0.12.0
llama-index-embeddings-huggingface==0.4.0
llama-index-embeddings-openai==0.3.1
llama-index-readers-file==0.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
openai==1.55.3
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
phidata==2.5.33
pillow==11.0.0
preshed==3.0.9
propcache==0.2.0
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
PyYAML==6.0.2
redis==5.0.1
regex==2024.9.11
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
rlp==4.0.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
starlette==0.41.3
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tomli==2.2.1
toolz==0.12.1
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typer==0.14.0
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.0
weasel==0.4.1
web3==7.2.0
websockets==13.0.1
wrapt==1.17.0
yarl==1.11.1
